core/requirements.txt
pyyaml
django
pandas
vobject
openpyxl
gunicorn
django-heroku
django-sortedm2m